Add frontend LaravelCmsPageRepository
